requirements.txt
akshare
google-genai
tavily-python
python-dotenv
pandas
openai
apscheduler
fastapi
uvicorn
yfinance
mplfinance
matplotlib
python-jose[cryptography]
passlib[argon2]
python-multipart
argon2-cffi
redis